requirements.txt
azure-cognitiveservices-speech==1.26.0
Flask==2.2.3
